Fix ImageView animations in LayoutTransition.
LayoutTransition animates left,top,right,bottom separately, which does
not cause setFrame() to be called at all; only onSizeChanged() is
triggered in that case. Because of that, the draw matrix isn't
touched and the animation seemingly doesn't work. As setFrame() also
calls onSizeChanged() if width or height changed and as computeBounds()
is only interested in width and height, but not in left and right, use
onSizeChanged() to call computeBounds() and thus update the draw matrix.

Change-Id: I7990b1677124c4aadd6e6baee9c36b3154c33d71